Fix publishing of newer drawing modules
Rename generated api bindings, which are not the
initial version of this API bindings.
This allows us to easily exclude those files from
publishing.
We still want to keep the full generated files in our
VCS since it allows us to track changes, and update
the modules containing the newly added apis for
a given feature level.

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>
Signed-off-by: Jonathan Schwender <jonathan.schwender@huawei.com>